+++ source/extensions/config_subscription/grpc/BUILD
@@ -14,6 +14,8 @@ envoy_cc_library(
     srcs = ["grpc_mux_impl.cc"],
     hdrs = ["grpc_mux_impl.h"],
     deps = [
+        ":grpc_stream_lib",
+        ":xds_source_id_lib",
         "//envoy/config:custom_config_validators_interface",
         "//envoy/config:grpc_mux_interface",
         "//envoy/config:subscription_interface",
@@ -25,12 +27,10 @@ envoy_cc_library(
         "//source/common/common:utility_lib",
         "//source/common/config:api_version_lib",
         "//source/common/config:decoded_resource_lib",
-        "//source/common/config:grpc_stream_lib",
         "//source/common/config:ttl_lib",
         "//source/common/config:utility_lib",
         "//source/common/config:xds_context_params_lib",
         "//source/common/config:xds_resource_lib",
-        "//source/common/config:xds_source_id_lib",
         "//source/common/memory:utils_lib",
         "//source/common/protobuf",
         "@com_google_absl//absl/container:btree",
@@ -44,13 +44,13 @@ envoy_cc_library(
     hdrs = ["new_grpc_mux_impl.h"],
     deps = [
         ":delta_subscription_state_lib",
+        ":grpc_stream_lib",
+        ":pausable_ack_queue_lib",
+        ":watch_map_lib",
         "//envoy/config:custom_config_validators_interface",
         "//envoy/config:xds_config_tracker_interface",
         "//envoy/event:dispatcher_interface",
         "//envoy/grpc:async_client_interface",
-        "//source/common/config:grpc_stream_lib",
-        "//source/common/config:pausable_ack_queue_lib",
-        "//source/common/config:watch_map_lib",
         "//source/common/config:xds_context_params_lib",
         "//source/common/config:xds_resource_lib",
         "//source/common/memory:utils_lib",
@@ -138,6 +138,8 @@ envoy_cc_library(
         "old_delta_subscription_state.h",
     ],
     deps = [
+        ":pausable_ack_queue_lib",
+        ":watch_map_lib",
         "//envoy/config:subscription_interface",
         "//envoy/event:dispatcher_interface",
         "//source/common/common:assert_lib",
@@ -145,11 +147,71 @@ envoy_cc_library(
         "//source/common/common:minimal_logger_lib",
         "//source/common/common:token_bucket_impl_lib",
         "//source/common/config:api_version_lib",
-        "//source/common/config:pausable_ack_queue_lib",
         "//source/common/config:ttl_lib",
         "//source/common/config:utility_lib",
-        "//source/common/config:watch_map_lib",
         "//source/common/grpc:common_lib",
+        "//source/common/protobuf",
+        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "grpc_stream_lib",
+    hdrs = ["grpc_stream.h"],
+    deps = [
+        "//envoy/config:grpc_mux_interface",
+        "//envoy/config:subscription_interface",
+        "//envoy/grpc:async_client_interface",
+        "//envoy/upstream:cluster_manager_interface",
+        "//source/common/common:backoff_lib",
+        "//source/common/common:minimal_logger_lib",
+        "//source/common/common:token_bucket_impl_lib",
+        "//source/common/config:utility_lib",
+        "//source/common/grpc:async_client_lib",
+        "//source/common/protobuf",
+    ],
+)
+
+envoy_cc_library(
+    name = "pausable_ack_queue_lib",
+    srcs = ["pausable_ack_queue.cc"],
+    hdrs = ["pausable_ack_queue.h"],
+    deps = [
+        ":update_ack_lib",
+        "//source/common/common:assert_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "xds_source_id_lib",
+    srcs = ["xds_source_id.cc"],
+    hdrs = ["xds_source_id.h"],
+    deps = [
+        "//envoy/config:xds_resources_delegate_interface",
+    ],
+)
+
+envoy_cc_library(
+    name = "update_ack_lib",
+    hdrs = ["update_ack.h"],
+    deps = ["@com_google_googleapis//google/rpc:status_cc_proto"],
+)
+
+envoy_cc_library(
+    name = "watch_map_lib",
+    srcs = ["watch_map.cc"],
+    hdrs = ["watch_map.h"],
+    deps = [
+        "//envoy/config:custom_config_validators_interface",
+        "//envoy/config:subscription_interface",
+        "//envoy/config:xds_config_tracker_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:cleanup_lib",
+        "//source/common/common:minimal_logger_lib",
+        "//source/common/common:utility_lib",
+        "//source/common/config:decoded_resource_lib",
+        "//source/common/config:utility_lib",
+        "//source/common/config:xds_resource_lib",
         "//source/common/protobuf",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
     ],
