+++ source/extensions/config_subscription/grpc/xds_mux/BUILD
@@ -32,9 +32,9 @@ envoy_cc_library(
         "//source/common/config:api_version_lib",
         "//source/common/config:decoded_resource_lib",
         "//source/common/config:utility_lib",
-        "//source/common/config:xds_source_id_lib",
         "//source/common/grpc:common_lib",
         "//source/common/protobuf",
+        "//source/extensions/config_subscription/grpc:xds_source_id_lib",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
     ],
 )
@@ -50,8 +50,8 @@ envoy_cc_library(
         "//envoy/local_info:local_info_interface",
         "//source/common/common:minimal_logger_lib",
         "//source/common/config:ttl_lib",
-        "//source/common/config:update_ack_lib",
         "//source/common/config:utility_lib",
+        "//source/extensions/config_subscription/grpc:update_ack_lib",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
     ],
 )
@@ -65,12 +65,12 @@ envoy_cc_extension(
         ":sotw_subscription_state_lib",
         "//envoy/event:dispatcher_interface",
         "//envoy/grpc:async_client_interface",
-        "//source/common/config:grpc_stream_lib",
-        "//source/common/config:pausable_ack_queue_lib",
-        "//source/common/config:watch_map_lib",
         "//source/common/config:xds_context_params_lib",
         "//source/common/config:xds_resource_lib",
         "//source/common/memory:utils_lib",
+        "//source/extensions/config_subscription/grpc:grpc_stream_lib",
+        "//source/extensions/config_subscription/grpc:pausable_ack_queue_lib",
+        "//source/extensions/config_subscription/grpc:watch_map_lib",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
     ],
 )
