+++ source/extensions/config_subscription/grpc/xds_mux/sotw_subscription_state.cc
@@ -1,7 +1,7 @@
 #include "source/extensions/config_subscription/grpc/xds_mux/sotw_subscription_state.h"
 
 #include "source/common/config/utility.h"
-#include "source/common/config/xds_source_id.h"
+#include "source/extensions/config_subscription/grpc/xds_source_id.h"
 
 namespace Envoy {
 namespace Config {
