+++ source/extensions/config_subscription/grpc/xds_source_id.cc
@@ -1,4 +1,4 @@
-#include "source/common/config/xds_source_id.h"
+#include "source/extensions/config_subscription/grpc/xds_source_id.h"
 
 #include "absl/strings/str_cat.h"
